refactoring mock prices location
