+++ prisma/seeds/prices.ts
@@ -2,6 +2,7 @@ import { Price, Prisma } from '@prisma/client'
 import { getAllPricesByNetworkTicker } from 'services/priceService'
 import { NETWORK_TICKERS, NETWORK_IDS, QUOTE_IDS, KeyValueT } from 'constants/index'
 import { readCsv, fileExists, isEmpty } from 'utils/index'
+import { mockPrices } from 'tests/mockedObjects'
 import moment from 'moment'
 import * as fs from 'fs'
 import * as path from 'path'
@@ -13,45 +14,6 @@ interface PriceFileData extends KeyValueT<string> {
   priceInCAD: string
   priceInUSD: string
 }
-
-export const mockPrices: Price[] = [
-  { // XECUSD
-    id: 0,
-    value: new Prisma.Decimal('0.00004095'),
-    createdAt: new Date(),
-    updatedAt: new Date(),
-    timestamp: 1664593200,
-    networkId: 1,
-    quoteId: 1
-  },
-  { // XECCAD
-    id: 0,
-    value: new Prisma.Decimal('0.00005663'),
-    createdAt: new Date(),
-    updatedAt: new Date(),
-    timestamp: 1664593200,
-    networkId: 1,
-    quoteId: 2
-  },
-  { // BCHUSD
-    id: 0,
-    value: new Prisma.Decimal('117.9081'),
-    createdAt: new Date(),
-    updatedAt: new Date(),
-    timestamp: 1664593200,
-    networkId: 2,
-    quoteId: 1
-  },
-  { // BCHCAD
-    id: 0,
-    value: new Prisma.Decimal('163.0735'),
-    createdAt: new Date(),
-    updatedAt: new Date(),
-    timestamp: 1664593200,
-    networkId: 2,
-    quoteId: 2
-  }
-]
 
 export const PATH_PRICE_CSV_FILE = path.join('prisma', 'seeds', 'prices.csv')
